backend/requirements.txt
fastapi
uvicorn[standard]
httpx
dnspython
python-whois
python-dotenv
slowapi
google-generativeai
